requirements.txt
certifi==2025.10.5
charset-normalizer==3.4.3
idna==3.10
PyPDF2==3.0.1
PyYAML==6.0.3
requests==2.32.5
urllib3==2.5.0
